requirements.txt
pytest==7.4.0
selenium==4.16.0
allure-pytest==2.13.2
PyYAML==6.0.1
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
    